# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SotkaCRM REST API & Gateway",
    "version": "1.0.0",
    "description": "Формализованная спецификация REST API SotkaCRM (v1) и интеграционного шлюза Sotka HQ. Охватывает авторизацию, управление лидами (с инвариантом soft-delete/отмены), базой продавцов, подключениями, выплатами сотрудникам, тарифами, ставками и аналитикой."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Локальный сервер разработки"
    },
    {
      "url": "https://sotkacrm.vercel.app",
      "description": "Production среда (Vercel)"
    }
  ],
...
  "paths": {
...
    "/api/sync/sotka": {
      "get": {
        "tags": [
          "Sync Gateway"
        ],
        "summary": "Проверка статуса шлюза синхронизации",
        "responses": {
          "200": {
            "description": "Шлюз синхронизации активен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sync Gateway"
        ],
        "summary": "Вебхук синхронизации данных из Sotka",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncWebhookPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Событие обработано"
          },
          "429": {
            "description": "Превышен лимит входящих вебхуков (максимум 30 запросов в минуту)",
            "headers": {
              "Retry-After": {
                "$ref": "#/components/headers/Retry-After"
              },
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/X-RateLimit-Limit"
              },
              "X-RateLimit-Remaining": {
                "$ref": "#/components/headers/X-RateLimit-Remaining"
              },
              "X-RateLimit-Reset": {
                "$ref": "#/components/headers/X-RateLimit-Reset"
              },
              "Server-Timing": {
                "$ref": "#/components/headers/Server-Timing"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "headers": {
      "Server-Timing": {
        "description": "Метрики задержки обработки запроса сервером (W3C Server-Timing)",
        "schema": {
          "type": "string",
          "example": "app;dur=45.2;desc=\"Application Processing\", total;dur=48.6"
        }
      },
      "X-RateLimit-Limit": {
        "description": "Максимальное количество разрешенных запросов в текущем окне",
        "schema": {
          "type": "integer",
          "example": 5
        }
      },
      "X-RateLimit-Remaining": {
        "description": "Оставшееся количество разрешенных запросов в окне",
        "schema": {
          "type": "integer",
          "example": 0
        }
      },
      "X-RateLimit-Reset": {
        "description": "Unix-метка времени (в секундах), когда окно лимита будет сброшено",
        "schema": {
          "type": "integer",
          "example": 1790510400
        }
      },
      "Retry-After": {
        "description": "Количество секунд, которые клиенту необходимо подождать перед повторным запросом",
        "schema": {
          "type": "integer",
          "example": 60
        }
      }
    },
    "schemas": {
      "RateLimitErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Слишком много запросов. Пожалуйста, подождите перед повторной попыткой."
          },
          "code": {
            "type": "string",
            "example": "RATE_LIMIT_EXCEEDED"
          },
          "retry_after": {
            "type": "integer",
            "example": 60
          }
        }
      },
...
      "SyncWebhookPayload": {
        "type": "object",
        "required": [
          "event",
          "data"
        ],
        "properties": {
          "event": {
            "type": "string",
            "enum": [
              "seller.created",
              "seller.updated",
              "lead.assigned",
              "moderation.changed"
            ],
            "example": "seller.updated"
          },
          "data": {
            "type": "object",
            "properties": {
              "organization_id": {
                "type": "integer",
                "example": 451
              },
              "status": {
                "type": "string",
                "example": "approved"
              },
              "comment": {
                "type": "string",
                "example": "Документы подтверждены"
              }
            }
          }
        }
      },
...
    }
  }
}